# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Role-Based Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export type UserProfileStatus = typeof UserProfileStatus[keyof typeof UserProfileStatus];


export const UserProfileStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UserProfile {
  id: string;
  clerkId: string;
  fullName: string;
  email: string;
  role: UserProfileRole;
  status: UserProfileStatus;
  isActive: boolean;
  createdAt: string;
}

export interface UserSync {
  fullName: string;
  email: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export type UserUpdateStatus = typeof UserUpdateStatus[keyof typeof UserUpdateStatus];


export const UserUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UserUpdate {
  fullName?: string;
  isActive?: boolean;
  role?: UserUpdateRole;
  status?: UserUpdateStatus;
}

export type TaskDurationUnit = typeof TaskDurationUnit[keyof typeof TaskDurationUnit];


export const TaskDurationUnit = {
  hours: 'hours',
  days: 'days',
  weeks: 'weeks',
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  incomplete: 'incomplete',
} as const;

export interface Task {
  id: string;
  title: string;
  description: string;
  assignedTo: string;
  assignedBy: string;
  /** @nullable */
  assignedToName?: string | null;
  /** @nullable */
  assignedByName?: string | null;
  durationValue: number;
  durationUnit: TaskDurationUnit;
  status: TaskStatus;
  createdAt: string;
  deadline: string;
  /** @nullable */
  completedAt?: string | null;
}

export type TaskInputDurationUnit = typeof TaskInputDurationUnit[keyof typeof TaskInputDurationUnit];


export const TaskInputDurationUnit = {
  hours: 'hours',
  days: 'days',
  weeks: 'weeks',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description: string;
  assignedTo: string;
  /** @minimum 1 */
  durationValue: number;
  durationUnit: TaskInputDurationUnit;
}

export type TaskEditDurationUnit = typeof TaskEditDurationUnit[keyof typeof TaskEditDurationUnit];


export const TaskEditDurationUnit = {
  hours: 'hours',
  days: 'days',
  weeks: 'weeks',
} as const;

export interface TaskEdit {
  /** @minLength 1 */
  title?: string;
  description?: string;
  assignedTo?: string;
  /** @minimum 1 */
  durationValue?: number;
  durationUnit?: TaskEditDurationUnit;
}

export type TaskStatusUpdateStatus = typeof TaskStatusUpdateStatus[keyof typeof TaskStatusUpdateStatus];


export const TaskStatusUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  incomplete: 'incomplete',
} as const;

export interface TaskStatusUpdate {
  status: TaskStatusUpdateStatus;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  present: 'present',
  absent: 'absent',
  late: 'late',
} as const;

export interface AttendanceRecord {
  id: string;
  employeeId: string;
  /** @nullable */
  employeeName?: string | null;
  date: string;
  checkIn: string;
  /** @nullable */
  checkOut?: string | null;
  status: AttendanceRecordStatus;
}

export interface DashboardStats {
  totalTasks: number;
  pendingTasks: number;
  completedTasks: number;
  incompleteTasks: number;
  inProgressTasks?: number;
  totalEmployees: number;
  presentToday: number;
  overdueTasksCount: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface AttendanceDaySummary {
  date: string;
  present: number;
  absent: number;
  late: number;
  total: number;
}