# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Role-Based Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from './userProfileRole';
import type { UserProfileStatus } from './userProfileStatus';

export interface UserProfile {
  id: string;
  clerkId: string;
  fullName: string;
  email: string;
  role: UserProfileRole;
  status: UserProfileStatus;
  isActive: boolean;
  createdAt: string;
}